move files and small refactoring
