+++ Markdown/Parsers/Configurators/TokenParserConfig.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Runtime.InteropServices.ComTypes;
 
 namespace Markdown
 {
