+++ Markdown/Parsers/Configurators/TokenRulesConfigurator.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace Markdown
+﻿namespace Markdown
 {
     internal class TokenRulesConfigurator
     {
